Remove Summary Agent and make Question Agent use RAG
