+++ engine/src/stirling/agents/pdf_questions.py
@@ -1,74 +1,94 @@
 from __future__ import annotations
 
+import logging
+
 from pydantic_ai import Agent
 from pydantic_ai.output import NativeOutput
 
-from stirling.agents._page_text import format_page_text, has_page_text
 from stirling.contracts import (
-    NeedContentFileRequest,
-    NeedContentResponse,
+    AiFile,
+    NeedIngestResponse,
     PdfContentType,
     PdfQuestionAnswerResponse,
     PdfQuestionNotFoundResponse,
     PdfQuestionRequest,
     PdfQuestionResponse,
+    PdfQuestionTerminalResponse,
     SupportedCapability,
     format_conversation_history,
     format_file_names,
 )
+from stirling.rag import RagCapability
 from stirling.services import AppRuntime
+
+logger = logging.getLogger(__name__)
+
+
+PDF_QUESTION_SYSTEM_PROMPT = (
+    "You answer questions about PDF documents by retrieving relevant content with the "
+    "search_knowledge tool. Use it before answering. Do not guess or use outside knowledge.\n"
+    "\n"
+    "Guidelines:\n"
+    "- Make one or two targeted search_knowledge calls based on the user's question. "
+    "Only search more if you genuinely need more context.\n"
+    "- Answer from the retrieved text. If the text doesn't support an answer, return not_found.\n"
+    "- Include a short list of evidence snippets (with page numbers where available) drawn "
+    "from what search_knowledge returned."
+)
 
 
 class PdfQuestionAgent:
     def __init__(self, runtime: AppRuntime) -> None:
         self.runtime = runtime
-        rag = runtime.rag_capability
-        self.agent = Agent(
-            model=runtime.smart_model,
-            output_type=NativeOutput(
-                [
-                    PdfQuestionAnswerResponse,
-                    PdfQuestionNotFoundResponse,
-                ]
-            ),
-            system_prompt=(
-                "Answer questions about PDFs using only the extracted page text provided in the prompt. "
-                "Do not guess or use outside knowledge. "
-                "If the answer is not supported by the provided text, return not_found. "
-                "When answering, include a short list of evidence snippets with their page numbers."
-            ),
-            instructions=rag.instructions,
-            toolsets=[rag.toolset],
-            model_settings=runtime.smart_model_settings,
-        )
 
     async def handle(self, request: PdfQuestionRequest) -> PdfQuestionResponse:
-        if not has_page_text(request.page_text):
-            return NeedContentResponse(
+        logger.info(
+            "[pdf-question] handle: files=%s question=%r",
+            [file.name for file in request.files],
+            request.question,
+        )
+        missing = await self._find_missing_files(request.files)
+        if missing:
+            logger.info("[pdf-question] missing ingestions: %s", [file.name for file in missing])
+            return NeedIngestResponse(
                 resume_with=SupportedCapability.PDF_QUESTION,
-                reason="No extracted PDF page text was provided, so the question cannot be answered yet.",
-                files=[
-                    NeedContentFileRequest(
-                        file_name=file.name,
-                        content_types=[PdfContentType.PAGE_TEXT],
-                    )
-                    for file in request.files
-                ],
-                max_pages=self.runtime.settings.max_pages,
-                max_characters=self.runtime.settings.max_characters,
+                reason="Some files have not been ingested into RAG yet.",
+                files_to_ingest=missing,
+                content_types=[PdfContentType.PAGE_TEXT],
             )
         return await self._run_answer_agent(request)
 
-    async def _run_answer_agent(self, request: PdfQuestionRequest) -> PdfQuestionResponse:
-        result = await self.agent.run(self._build_prompt(request))
+    async def _find_missing_files(self, files: list[AiFile]) -> list[AiFile]:
+        missing: list[AiFile] = []
+        for file in files:
+            if not await self.runtime.rag_service.has_collection(file.id):
+                missing.append(file)
+        return missing
+
+    async def _run_answer_agent(self, request: PdfQuestionRequest) -> PdfQuestionTerminalResponse:
+        rag = RagCapability(
+            rag_service=self.runtime.rag_service,
+            collections=[file.id for file in request.files],
+            top_k=self.runtime.settings.rag_default_top_k,
+        )
+        agent = Agent(
+            model=self.runtime.smart_model,
+            output_type=NativeOutput([PdfQuestionAnswerResponse, PdfQuestionNotFoundResponse]),
+            system_prompt=PDF_QUESTION_SYSTEM_PROMPT,
+            instructions=rag.instructions,
+            toolsets=[rag.toolset],
+            model_settings=self.runtime.smart_model_settings,
+        )
+        prompt = self._build_prompt(request)
+        logger.debug("[pdf-question] prompt:\n%s", prompt)
+        result = await agent.run(prompt)
         return result.output
 
     def _build_prompt(self, request: PdfQuestionRequest) -> str:
-        pages = format_page_text(request.page_text, empty="")
         history = format_conversation_history(request.conversation_history)
         return (
             f"Conversation history:\n{history}\n"
             f"Files: {format_file_names(request.files)}\n"
             f"Question: {request.question}\n"
-            f"Extracted page text:\n{pages}"
+            "Use search_knowledge to retrieve the relevant content, then answer."
         )
