+++ engine/tests/test_stirling_api.py
@@ -7,7 +7,6 @@ from stirling.api.dependencies import (
     get_orchestrator_agent,
     get_pdf_edit_agent,
     get_pdf_question_agent,
-    get_summary_agent,
     get_user_spec_agent,
 )
 from stirling.config import load_settings
@@ -21,12 +20,10 @@ from stirling.contracts import (
     CannotContinueExecutionAction,
     EditCannotDoResponse,
     NeedContentResponse,
-    NeedIngestResponse,
     OrchestratorRequest,
     PdfEditRequest,
     PdfQuestionNotFoundResponse,
     PdfQuestionRequest,
-    SummaryRequest,
     SupportedCapability,
 )
 from stirling.models.tool_models import Angle, RotatePdfParams
@@ -53,15 +50,6 @@ class StubPdfQuestionAgent:
         return PdfQuestionNotFoundResponse(reason=request.question)
 
 
-class StubSummaryAgent:
-    async def handle(self, request: SummaryRequest) -> NeedIngestResponse:
-        return NeedIngestResponse(
-            resume_with=SupportedCapability.PDF_SUMMARY,
-            reason="stub-not-ingested",
-            files_to_ingest=request.files,
-        )
-
-
 class StubUserSpecAgent:
     async def draft(self, request: AgentDraftRequest) -> AgentDraftResponse:
         return AgentDraftResponse(
@@ -86,7 +74,6 @@ app.dependency_overrides[load_settings] = build_app_settings
 app.dependency_overrides[get_orchestrator_agent] = lambda: StubOrchestratorAgent()
 app.dependency_overrides[get_pdf_edit_agent] = lambda: StubPdfEditAgent()
 app.dependency_overrides[get_pdf_question_agent] = lambda: StubPdfQuestionAgent()
-app.dependency_overrides[get_summary_agent] = lambda: StubSummaryAgent()
 app.dependency_overrides[get_user_spec_agent] = lambda: StubUserSpecAgent()
 app.dependency_overrides[get_execution_planning_agent] = lambda: StubExecutionPlanningAgent()
 
@@ -123,29 +110,11 @@ def test_pdf_questions_route() -> None:
         json={
             "question": "what is this?",
             "files": [{"id": "test-id", "name": "test.pdf"}],
-            "pageText": [{"fileName": "test.pdf", "pages": [{"pageNumber": 1, "text": "Example"}]}],
         },
     )
 
     assert response.status_code == 200
     assert response.json()["outcome"] == "not_found"
-
-
-def test_pdf_summary_route() -> None:
-    response = client.post(
-        "/api/v1/pdf/summary",
-        json={
-            "files": [
-                {"id": "doc-1-id", "name": "doc-1.pdf"},
-                {"id": "doc-2-id", "name": "doc-2.pdf"},
-            ],
-        },
-    )
-
-    assert response.status_code == 200
-    body = response.json()
-    assert body["outcome"] == "need_ingest"
-    assert [file["id"] for file in body["filesToIngest"]] == ["doc-1-id", "doc-2-id"]
 
 
 def test_agent_draft_route() -> None:
