+++ frontend/src/prototypes/components/chat/ChatPanel.tsx
@@ -29,7 +29,6 @@ import ExpandLessIcon from "@mui/icons-material/ExpandLess";
 import {
   useChat,
   AiWorkflowPhase,
-  type AiSummaryResult,
   type AiWorkflowProgress,
 } from "@app/components/chat/ChatContext";
 import { useTranslatedToolCatalog } from "@app/data/useTranslatedToolRegistry";
@@ -142,62 +141,25 @@ function ToolsUsedBlock({
   );
 }
 
-function SummaryBlock({ summary }: { summary: AiSummaryResult }) {
-  return (
-    <Stack gap="xs">
-      <Text size="sm" fw={600}>
-        {summary.tldr}
-      </Text>
-      {summary.keyPoints.length > 0 && (
-        <List size="sm" spacing={2} pl="sm">
-          {summary.keyPoints.map((point, i) => (
-            <List.Item key={i}>{point}</List.Item>
-          ))}
-        </List>
-      )}
-      {summary.sections.length > 0 && (
-        <Stack gap={4} mt="xs">
-          {summary.sections.map((section, i) => (
-            <Box key={i}>
-              <Text size="sm" fw={600}>
-                {section.heading}
-              </Text>
-              <Text size="sm" style={{ whiteSpace: "pre-wrap" }}>
-                {section.summary}
-              </Text>
-            </Box>
-          ))}
-        </Stack>
-      )}
-    </Stack>
-  );
-}
-
 function ChatMessageBubble({
   role,
   content,
   toolsUsed,
-  summary,
   resolveToolName,
   t,
 }: {
   role: "user" | "assistant";
   content: string;
   toolsUsed?: string[];
-  summary?: AiSummaryResult;
   resolveToolName: ToolNameResolver;
   t: TranslateFn;
 }) {
   return (
     <div className={`chat-message chat-message-${role}`}>
       <Paper className={`chat-bubble chat-bubble-${role}`} p="xs" radius="md">
-        {summary ? (
-          <SummaryBlock summary={summary} />
-        ) : (
-          <Text size="sm" style={{ whiteSpace: "pre-wrap" }}>
-            {content}
-          </Text>
-        )}
+        <Text size="sm" style={{ whiteSpace: "pre-wrap" }}>
+          {content}
+        </Text>
         {toolsUsed && toolsUsed.length > 0 && (
           <ToolsUsedBlock
             tools={toolsUsed}
@@ -299,7 +261,6 @@ export function ChatPanel() {
                     role={msg.role}
                     content={msg.content}
                     toolsUsed={msg.toolsUsed}
-                    summary={msg.summary}
                     resolveToolName={resolveToolName}
                     t={t}
                   />
